refactor(v2): extract card binder and pack flow (#234)
